+++ src/core/spec/launchApp.spec.js
@@ -90,6 +90,7 @@ describe('launchApp() verify execution order of life cycle hooks (both Aspects a
     { type: 'alch', func: 'initialRootAppElm' },
     { type: 'flch', func: 'appWillStart' },
     { type: 'alch', func: 'injectRootAppElm' },
+    { type: 'alch', func: 'injectParamsInHooks' },
     { type: 'flch', func: 'appInit' },
     { type: 'flch', func: 'appDidStart' },
   ];